+++ tests/api-resources/servers/icon.test.ts
@@ -7,10 +7,22 @@ const client = new Smithery({
   baseURL: process.env['TEST_API_BASE_URL'] ?? 'http://127.0.0.1:4010',
 });
 
-describe('resource uplink', () => {
+describe('resource icon', () => {
   // Mock server tests are disabled
-  test.skip('createToken', async () => {
-    const responsePromise = client.uplink.createToken();
+  test.skip('delete', async () => {
+    const responsePromise = client.servers.icon.delete('qualifiedName');
+    const rawResponse = await responsePromise.asResponse();
+    expect(rawResponse).toBeInstanceOf(Response);
+    const response = await responsePromise;
+    expect(response).not.toBeInstanceOf(Response);
+    const dataAndResponse = await responsePromise.withResponse();
+    expect(dataAndResponse.data).toBe(response);
+    expect(dataAndResponse.response).toBe(rawResponse);
+  });
+
+  // Mock server tests are disabled
+  test.skip('upload', async () => {
+    const responsePromise = client.servers.icon.upload('qualifiedName');
     const rawResponse = await responsePromise.asResponse();
     expect(rawResponse).toBeInstanceOf(Response);
     const response = await responsePromise;