+++ tests/api-resources/servers/repo.test.ts
@@ -8,7 +8,35 @@ const client = new Smithery({
 });
 
 describe('resource repo', () => {
-  // Prism tests are disabled
+  // Mock server tests are disabled
+  test.skip('update', async () => {
+    const responsePromise = client.servers.repo.update('qualifiedName');
+    const rawResponse = await responsePromise.asResponse();
+    expect(rawResponse).toBeInstanceOf(Response);
+    const response = await responsePromise;
+    expect(response).not.toBeInstanceOf(Response);
+    const dataAndResponse = await responsePromise.withResponse();
+    expect(dataAndResponse.data).toBe(response);
+    expect(dataAndResponse.response).toBe(rawResponse);
+  });
+
+  // Mock server tests are disabled
+  test.skip('update: request options and params are passed correctly', async () => {
+    // ensure the request options are being passed correctly by passing an invalid HTTP method in order to cause an error
+    await expect(
+      client.servers.repo.update(
+        'qualifiedName',
+        {
+          autoDeploy: true,
+          baseDirectory: 'x',
+          branch: 'branch',
+        },
+        { path: '/_stainless_unknown_path' },
+      ),
+    ).rejects.toThrow(Smithery.NotFoundError);
+  });
+
+  // Mock server tests are disabled
   test.skip('delete', async () => {
     const responsePromise = client.servers.repo.delete('qualifiedName');
     const rawResponse = await responsePromise.asResponse();
@@ -20,7 +48,7 @@ describe('resource repo', () => {
     expect(dataAndResponse.response).toBe(rawResponse);
   });
 
-  // Prism tests are disabled
+  // Mock server tests are disabled
   test.skip('get', async () => {
     const responsePromise = client.servers.repo.get('qualifiedName');
     const rawResponse = await responsePromise.asResponse();
@@ -32,7 +60,7 @@ describe('resource repo', () => {
     expect(dataAndResponse.response).toBe(rawResponse);
   });
 
-  // Prism tests are disabled
+  // Mock server tests are disabled
   test.skip('set: only required params', async () => {
     const responsePromise = client.servers.repo.set('qualifiedName', { repoName: 'x', repoOwner: 'x' });
     const rawResponse = await responsePromise.asResponse();
@@ -44,7 +72,7 @@ describe('resource repo', () => {
     expect(dataAndResponse.response).toBe(rawResponse);
   });
 
-  // Prism tests are disabled
+  // Mock server tests are disabled
   test.skip('set: required and optional params', async () => {
     const response = await client.servers.repo.set('qualifiedName', {
       repoName: 'x',